+++ driver/aws/dynamoset/set.go
@@ -9,6 +9,8 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
 	"github.com/dogmatiq/persistencekit/driver/aws/internal/awsx"
+	"github.com/dogmatiq/persistencekit/driver/aws/internal/dynamox"
+	"github.com/dogmatiq/persistencekit/set"
 )
 
 type setimpl struct {
@@ -16,12 +18,13 @@ type setimpl struct {
 	OnRequest func(any) []func(*dynamodb.Options)
 
 	attr struct {
-		Set   types.AttributeValueMemberS
-		Value types.AttributeValueMemberB
+		Set    types.AttributeValueMemberS
+		Member types.AttributeValueMemberB
 	}
 
 	request struct {
 		Has    dynamodb.GetItemInput
+		Range  dynamodb.QueryInput
 		Put    dynamodb.PutItemInput
 		Delete dynamodb.DeleteItemInput
 	}
@@ -32,7 +35,7 @@ func (s *setimpl) Name() string {
 }
 
 func (s *setimpl) Has(ctx context.Context, v []byte) (bool, error) {
-	s.attr.Value.Value = v
+	s.attr.Member.Value = v
 
 	out, err := awsx.Do(
 		ctx,
@@ -41,7 +44,7 @@ func (s *setimpl) Has(ctx context.Context, v []byte) (bool, error) {
 		&s.request.Has,
 	)
 	if err != nil {
-		return false, fmt.Errorf("unable to get set value: %w", err)
+		return false, fmt.Errorf("unable to get set member: %w", err)
 	}
 
 	return out.Item != nil, nil
@@ -53,8 +56,8 @@ func (s *setimpl) Add(ctx context.Context, v []byte) error {
 }
 
 var (
-	mustExist    = aws.String("attribute_exists(" + valueAttr + ")")
-	mustNotExist = aws.String("attribute_not_exists(" + valueAttr + ")")
+	mustExist    = aws.String("attribute_exists(" + memberAttr + ")")
+	mustNotExist = aws.String("attribute_not_exists(" + memberAttr + ")")
 )
 
 func (s *setimpl) TryAdd(ctx context.Context, v []byte) (bool, error) {
@@ -86,8 +89,29 @@ func (s *setimpl) TryRemove(ctx context.Context, v []byte) (bool, error) {
 	return true, err
 }
 
+func (s *setimpl) Range(ctx context.Context, fn set.BinaryRangeFunc) error {
+	if err := dynamox.QueryRange(
+		ctx,
+		s.Client,
+		s.OnRequest,
+		&s.request.Range,
+		func(ctx context.Context, item map[string]types.AttributeValue) (bool, error) {
+			value, err := dynamox.AsBytes(item, memberAttr)
+			if err != nil {
+				return false, err
+			}
+
+			return fn(ctx, value)
+		},
+	); err != nil {
+		return fmt.Errorf("unable to range over set: %w", err)
+	}
+
+	return nil
+}
+
 func (s *setimpl) add(ctx context.Context, v []byte) error {
-	s.attr.Value.Value = v
+	s.attr.Member.Value = v
 
 	if _, err := awsx.Do(
 		ctx,
@@ -95,14 +119,14 @@ func (s *setimpl) add(ctx context.Context, v []byte) error {
 		s.OnRequest,
 		&s.request.Put,
 	); err != nil {
-		return fmt.Errorf("unable to put set value: %w", err)
+		return fmt.Errorf("unable to put set member: %w", err)
 	}
 
 	return nil
 }
 
 func (s *setimpl) delete(ctx context.Context, v []byte) error {
-	s.attr.Value.Value = v
+	s.attr.Member.Value = v
 
 	if _, err := awsx.Do(
 		ctx,
@@ -110,7 +134,7 @@ func (s *setimpl) delete(ctx context.Context, v []byte) error {
 		s.OnRequest,
 		&s.request.Delete,
 	); err != nil {
-		return fmt.Errorf("unable to delete set value: %w", err)
+		return fmt.Errorf("unable to delete set member: %w", err)
 	}
 
 	return nil